Frode: Added folder 1D2D, containing fw_and_bw.py.
This is a modification to the 2D-scheme in the 2D folder. The file contains two functions, which are able to do explicit and implicit (i.e. forward and backward) solutions to the 1D and 2D problem. There are also two trial functions, meant to demonstrate the type of solutions one might get.